requirements.txt
PyQt5
python-dotenv
mysql-connector-python
ultralytics
opencv-python
SpeechRecognition
PyAudio